tests/wpt/tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/form-submit-button-click-during-pageshow.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<body>
<script>
"use strict";
promise_test(async t => {
  const sentinelIframe = await setupSentinelIframe(t);
  const startingHistoryLength = history.length;

  const code = `
    window.onpageshow = () => {
      const form = document.createElement("form");
      form.action = "/common/blank.html";

      const input = document.createElement("input");
      input.type = "hidden";
      input.name = "thereplacement";
      form.append(input);

      const button = document.createElement("button");
      button.type = "submit";
      form.append(button);

      document.body.append(form);
      button.click();
    };
  `;

  const startURL = "resources/code-injector.html?pipe=sub(none)&code=" + encodeURIComponent(code);
  const afterReplacementURL = "/common/blank.html?thereplacement=";
  const iframe = insertIframe(t, startURL);

  assert_equals(history.length, startingHistoryLength, "Inserting the under-test iframe must not change history.length");

  await waitForLoadAllowingIntermediateLoads(t, iframe, afterReplacementURL);
  assert_equals(history.length, startingHistoryLength, "history.length must not change after waiting for the replacement");

  await checkSentinelIframe(t, sentinelIframe);
  assert_equals(history.length, startingHistoryLength, "history.length must not change after checking the sentinel iframe");
}, "Replace during pageshow, triggered by submitButton.click()");
</script>
